chore: build the environment in only one place
